Adding aliases and ffmpeg

+++ welcome.sh
@@ -1,21 +1,21 @@
 #!/bin/bash
-bash homebrew_install.sh && bash preferences.sh
+# bash homebrew_install.sh && bash preferences.sh
 
 # Symlink dotfiles
-rm -i ~/.bash_profile
+[ -f ~/.bash_profile ] && rm -i ~/.bash_profile || echo "Creating .bash_profile"
 ln -s ~/code/dotfiles/bash_profile ~/.bash_profile
 
-rm -i ~/.git-completion.bash
-ln -s ~/code/dotfiles/git-completion.bash ~/.git-completion.bash
+[ -f ~/.git_completion.bash ] && rm -i ~/.git_completion.bash || echo "Creating .git_completion.bash"
+ln -s ~/code/dotfiles/git-completion ~/.git-completion.bash
 
-rm -i ~/.gitconfig
+[ -f ~/.gitconfig.bash ] && rm -i ~/.gitconfig.bash || echo "Creating .gitconfig"
 ln -s ~/code/dotfiles/gitconfig ~/.gitconfig
 
-rm -i ~/.gitignore
+[ -f ~/.gitignore ] && rm -i ~/.gitignore || echo "Creating .gitignore"
 ln -s ~/code/dotfiles/gitignore ~/.gitignore
 
-rm -i ~/.irbrc
+[ -f ~/.irbrc.bash ] && rm -i ~/.irbrc.bash || echo "Creating .irbrc"
 ln -s ~/code/dotfiles/irbrc ~/.irbrc
 
-rm -i ~/.aliases
+[ -f ~/.aliases.bash ] && rm -i ~/.aliases.bash || echo "Creating .aliases"
 ln -s ~/code/dotfiles/aliases ~/.aliases.bash
